Makefile: Add the ability to build with libasan
Add the ability to build with the Address Sanitizer[0] by doing

    $ make ASAN=1

For this to work you will need libasan installed. On Fedora & CentOS
this is simply the libasan package.

This will then do various memory checks at run time and print a bunch of
diagnostics if any problems are detected.

[0]: https://github.com/google/sanitizers/wiki/AddressSanitizer

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>